Rename and move messaging metrics tests
