+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/Messaging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aefaae24291850b4e803f5bb4311540a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
