+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/Messaging/NetworkMetricsMessagingDispatchTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b431a89698f05134cb528420e812855e
+guid: f6c78652c40d92542b3e05b2998956e8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
