+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/Messaging/NetworkMetricsMessagingReceptionTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b27007e2b6be28642be778a9dd0fac62
+guid: c2bed8d64ee78b3428f4b4e68e921755
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
